refactor(api): move saml routes from lib to src
